server/createAdventure/requirements.txt
asgiref==3.3.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-enumfield==2.0.2
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
